+++ sites/public/src/components/account/AccountCard.module.scss
@@ -1,0 +1,18 @@
+.account-card {
+  --card-header-padding-inline-desktop: var(--seeds-s12);
+  --card-header-padding-inline-mobile: var(--seeds-s4);
+  --card-content-padding-inline-mobile: var(--seeds-s4);
+  @media (max-width: theme("screens.md")) {
+    --card-spacing-lg: var(--seeds-s6);
+  }
+
+  width: 100%;
+}
+
+.acccount-card-icon {
+  margin-bottom: var(--seeds-s3);
+}
+.account-card-header {
+  --subheading-margin: var(--seeds-s3);
+  --heading-margin: var(--seeds-s3);
+}
