--- sites/public/src/pages/account/AccountCard.tsx
@@ -1,86 +1,0 @@
-import { HeadingGroup, Icon } from "@bloom-housing/ui-seeds"
-import Card from "@bloom-housing/ui-seeds/src/blocks/Card"
-import React from "react"
-import styles from "./account.module.scss"
-
-interface AccountCardProps {
-  iconSymbol: "application" | "profile"
-  title: string
-  subtitle: string
-  children: React.ReactElement
-  id?: string
-  divider?: "flush" | "inset"
-  headingPriority?: 1 | 2 | 3 | 4 | 5 | 6
-  className?: string
-}
-
-const AccountCard = (props: AccountCardProps) => {
-  const classNames = [styles["account-card"]]
-  if (props.className) classNames.push(props.className)
-
-  const applicationIcon = (
-    <svg xmlns="http://www.w3.org/2000/svg" width="37" height="48" viewBox="0 0 37 48" fill="none">
-      <path d="M36.7665 11.439L25.3275 0H0V47.388H36.765V11.439H36.7665Z" fill="#EDEBDA" />
-      <path d="M25.3286 0V11.439H36.7676L25.3286 0Z" fill="#CEC9AE" />
-      <path
-        d="M28.5975 17.9728H15.525C15.0735 17.9728 14.709 17.6083 14.709 17.1568C14.709 16.7053 15.0765 16.3408 15.525 16.3408H28.5975C29.049 16.3408 29.4135 16.7083 29.4135 17.1568C29.4135 17.6053 29.046 17.9728 28.5975 17.9728Z"
-        fill="#CEC9AE"
-      />
-      <path
-        d="M8.98976 18.7922C8.78126 18.7922 8.56976 18.7112 8.41076 18.5522L6.77576 16.9172C6.45626 16.5977 6.45626 16.0817 6.77576 15.7622C7.09526 15.4427 7.61126 15.4427 7.93076 15.7622L9.04676 16.8782L12.5598 14.0672C12.9123 13.7867 13.4268 13.8422 13.7088 14.1947C13.9908 14.5472 13.9338 15.0617 13.5813 15.3437L9.49676 18.6122C9.34526 18.7322 9.16526 18.7922 8.98526 18.7922H8.98976Z"
-        fill="#13A085"
-      />
-      <path
-        d="M28.5975 26.9611H15.525C15.0735 26.9611 14.709 26.5966 14.709 26.1451C14.709 25.6936 15.0765 25.3291 15.525 25.3291H28.5975C29.049 25.3291 29.4135 25.6936 29.4135 26.1451C29.4135 26.5966 29.046 26.9611 28.5975 26.9611Z"
-        fill="#CEC9AE"
-      />
-      <path
-        d="M8.98976 27.7771C8.78126 27.7771 8.56976 27.6976 8.41076 27.5371L6.77576 25.9021C6.45626 25.5826 6.45626 25.0666 6.77576 24.7471C7.09526 24.4276 7.61126 24.4276 7.93076 24.7471L9.04676 25.8631L12.5598 23.0521C12.9123 22.7716 13.4268 22.8271 13.7088 23.1796C13.9908 23.5321 13.9338 24.0466 13.5813 24.3286L9.49676 27.5971C9.34526 27.7171 9.16526 27.7771 8.98526 27.7771H8.98976Z"
-        fill="#13A085"
-      />
-      <path
-        d="M28.5975 35.9489H15.525C15.0735 35.9489 14.709 35.5844 14.709 35.1329C14.709 34.6814 15.0765 34.3169 15.525 34.3169H28.5975C29.049 34.3169 29.4135 34.6814 29.4135 35.1329C29.4135 35.5844 29.046 35.9489 28.5975 35.9489Z"
-        fill="#CEC9AE"
-      />
-      <path
-        d="M8.98976 36.7649C8.78126 36.7649 8.56976 36.6854 8.41076 36.5249L6.77576 34.8899C6.45626 34.5704 6.45626 34.0544 6.77576 33.7349C7.09526 33.4154 7.61126 33.4154 7.93076 33.7349L9.04676 34.8509L12.5598 32.0399C12.9123 31.7594 13.4268 31.8149 13.7088 32.1674C13.9908 32.5199 13.9338 33.0344 13.5813 33.3164L9.49676 36.5849C9.34526 36.7049 9.16526 36.7649 8.98526 36.7649H8.98976Z"
-        fill="#13A085"
-      />
-    </svg>
-  )
-
-  const profileIcon = (
-    <svg xmlns="http://www.w3.org/2000/svg" width="48" height="48" viewBox="0 0 48 48" fill="none">
-      <path
-        d="M16.8447 37.6683L9.68974 41.5631C9.23674 41.8346 8.87374 42.106 8.51074 42.379C12.6777 45.9108 18.1107 47.9939 23.9982 47.9939C29.8857 47.9939 35.2272 45.9108 39.3942 42.4704C39.0312 42.1075 38.5782 41.8376 38.1267 41.6546L30.4272 37.8513C29.4312 37.3984 28.7982 36.3111 28.7982 35.2268V32.2394C28.9782 31.9679 29.2512 31.6965 29.5227 31.335C30.6102 29.8863 31.3347 28.2561 31.8777 26.5359C32.8737 26.2645 33.5982 25.3572 33.5982 24.2729V21.1024C33.5982 20.3781 33.3267 19.7437 32.7822 19.2908V14.4917C32.7822 14.4917 33.7782 7.24805 23.9067 7.24805C14.0352 7.24805 15.0312 14.4917 15.0312 14.4917V19.1108C14.5782 19.5637 14.2152 20.1981 14.2152 20.9225V24.0929C14.2152 24.9087 14.6682 25.7216 15.3027 26.0845C16.1187 29.6164 18.2022 32.0624 18.2022 32.0624V34.9614C18.3822 36.1401 17.7492 37.1359 16.8432 37.6773L16.8447 37.6683Z"
-        fill="#E7ECED"
-      />
-      <path
-        d="M24.4534 0.0001395C11.1409 -0.179827 0.273363 10.3227 0.000362808 23.5427C-0.0911372 31.0578 3.25986 37.8485 8.51286 42.2862C8.87586 42.0148 9.23736 41.7433 9.69186 41.4704L16.8469 37.5756C17.7514 37.0327 18.3874 36.0354 18.3874 35.0396V32.1406C18.3874 32.1406 16.3039 29.6046 15.4879 26.1627C14.8549 25.7098 14.4004 24.984 14.4004 24.1711V21.0007C14.4004 20.2763 14.6719 19.642 15.2164 19.189V14.4814C15.2164 14.4814 14.2204 7.23778 24.0919 7.23778C33.9634 7.23778 32.9674 14.4814 32.9674 14.4814V19.1006C33.4204 19.5535 33.7834 20.1879 33.7834 20.9122V24.0826C33.7834 25.1699 33.0589 26.0742 32.0629 26.3457C31.5199 28.0659 30.7039 29.6961 29.7079 31.1448C29.4364 31.5077 29.1649 31.8692 28.9834 32.0491V35.0366C28.9834 36.1239 29.6164 37.1197 30.6124 37.6611L38.3119 41.4644C38.7649 41.7358 39.2164 42.0073 39.5794 42.2802C44.5594 38.114 47.9119 31.6862 48.0004 24.4426C48.1804 11.1326 37.6759 0.267089 24.4534 -0.00585938V0.0001395Z"
-        fill="#556080"
-      />
-    </svg>
-  )
-
-  const customIcon = props.iconSymbol === "application" ? applicationIcon : profileIcon
-
-  return (
-    <Card spacing="lg" className={classNames.join(" ")}>
-      <Card.Header divider={props?.divider}>
-        <Icon size="2xl" className={styles["account-card-icon"]}>
-          {customIcon}
-        </Icon>
-        <HeadingGroup
-          size="2xl"
-          heading={props.title}
-          subheading={props.subtitle}
-          className={styles["account-card-header"]}
-          headingPriority={props.headingPriority}
-        />
-      </Card.Header>
-      {props.children}
-    </Card>
-  )
-}
-
-export { AccountCard as default, AccountCard }
