+++ sites/public/src/pages/account/account.module.scss
@@ -6,25 +6,6 @@
   }
 }
 
-.account-card {
-  --card-header-padding-inline-desktop: var(--seeds-s12);
-  --card-header-padding-inline-mobile: var(--seeds-s4);
-  --card-content-padding-inline-mobile: var(--seeds-s4);
-  @media (max-width: theme("screens.md")) {
-    --card-spacing-lg: var(--seeds-s6);
-  }
-
-  width: 100%;
-}
-
-.acccount-card-icon {
-  margin-bottom: var(--seeds-s3);
-}
-.account-card-header {
-  --subheading-margin: var(--seeds-s3);
-  --heading-margin: var(--seeds-s3);
-}
-
 .account-card-section {
   @media (min-width: theme("screens.md")) {
     --card-content-padding-inline: var(--seeds-s12);
